# Optimized find pooled resources in ZonalParkingManager