@REASY 2018-10-10
I did run in cluster(master+2workers). Here is output for 2 iterations: https://drive.google.com/open?id=1yLVFVk_5q1O2wTOdx2TsfaysvDxPPfMz
Agent Sim times:
Iteration 0 => 5536 seconds
Iteration 1 => 3416 seconds.