@REASY 2018-10-11
Results:
No optimization: https://drive.google.com/open?id=105GaMdED7el7aqkrqK5FPYu6eN9tH_1d
With optimization: https://drive.google.com/open?id=1uOgd90tzWRYSyjbgBaoiBkz2KlmViFT6