@REASY 2018-10-11
Call tree for non-optimized version:
![image](https://user-images.githubusercontent.com/5107562/46821027-e1713b80-cdb1-11e8-849b-c1e6c1b23bde.png)