@REASY 2018-10-11
Call tree for optimized one:
![image](https://user-images.githubusercontent.com/5107562/46821202-4593ff80-cdb2-11e8-9404-a4ba984da938.png)