@REASY 2018-10-13
New run with 5 iterations
No optimization: https://drive.google.com/open?id=1dNUe6uqM55NmobqWIQyZv9bWP729os_X
Time:
Iteration 0 => 7169 s
Iteration 1 => 5153 s
Iteration 2 => 5202 s
Iteration 3 => 5350 s
Iteration 4 => 5138 s